+++ TestExtensions/user_phpunittest2/Configuration/TCA/user_phpunittest2_test.php
@@ -13,6 +13,7 @@ return [
             'endtime' => 'endtime',
         ],
         'hideTable' => true,
+        'adminOnly' => true,
     ],
     'columns' => [
         'hidden' => [