+++ Classes/TestRunner/CliTestRunner.php
@@ -20,7 +20,7 @@ class Tx_Phpunit_TestRunner_CliTestRunner extends \Tx_Phpunit_TestRunner_Abstrac
         // This is needed as the configuration might include closures which cannot be backed up
         $globalBackup = $this->removeClosures($GLOBALS['TYPO3_CONF_VARS']);
 
-        require_once __DIR__ . '/../../Resources/Private/Libraries/phpunit-library.phar';
+        require_once PATH_site . 'typo3conf/ext/phpunit/Resources/Private/Libraries/phpunit-library.phar';
 
         // Run unit tests
         /** @var string */