+++ Classes/Exception/Database.php
@@ -10,6 +10,8 @@ use TYPO3\CMS\Core\Exception;
  * The exception automatically will use an error message, the error message
  * from the DB and the last query.
  *
+ * @deprecated will be removed for PHPUnit 6.
+ *
  * @author Oliver Klee <typo3-coding@oliverklee.de>
  */
 class Tx_Phpunit_Exception_Database extends Exception