+++ Classes/Interface/FrameworkCleanupHook.php
@@ -4,6 +4,8 @@
  * This interface should be used for classes that should get called by a hook
  * when \Tx_Phpunit_Framework::cleanUp() is called.
  *
+ * @deprecated will be removed for PHPUnit 6.
+ *
  * @author Oliver Klee <typo3-coding@oliverklee.de>
  */
 interface Tx_Phpunit_Interface_FrameworkCleanupHook