+++ Classes/Service/ExtensionSettingsService.php
@@ -5,6 +5,8 @@ use TYPO3\CMS\Core\SingletonInterface;
 /**
  * This class provides functions for reading the settings of the PHPUnit extension (as set in the extension manager).
  *
+ * @deprecated will be removed for PHPUnit 6.
+ *
  * @author Oliver Klee <typo3-coding@oliverklee.de>
  */
 class Tx_Phpunit_Service_ExtensionSettingsService extends \Tx_Phpunit_AbstractDataContainer implements \Tx_Phpunit_Interface_ExtensionSettingsService, SingletonInterface