+++ Classes/BackEnd/index.php
@@ -1,5 +1,6 @@
 <?php
 
+use OliverKlee\Phpunit\Service\LibraryLoader;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
@@ -11,6 +12,8 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  * @author Oliver Klee <typo3-coding@oliverklee.de>
  */
 defined('TYPO3_MODE') or die('Access denied.');
+
+LibraryLoader::includeAll();
 
 /** @var \TYPO3\CMS\Lang\LanguageService $languageService */
 $languageService = $GLOBALS['LANG'];