+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Removed
+- Remove the unused PHPUnit configuration file (#76)
 - Remove the obsolete ext_autoload.php (#56)
 
 ### Fixed